addchain/acc: make Compile a pass
This feels a little unusual, but the intention is that the results of
Compile() and Eval() are useful for other passes or analyses, so it
makes sense to store them in the ir.Program. At this point they seem be
playing more of a role as "passes".

Extracted-from: mmcloughlin/ec3@4143a587c1c09e28995655b5ba3049de982d5bc9