alg/dict: SlidingWindow: max zeroes and shortening
This change adds new variations of SlidingWindow and modifies
Hybrid to use an arbitrary Decomposer after runs are removed.
SlidingWindowRTL and SlidingWindowShortRTL construct the windows
from least to most significant bit (right-to-left) instead of the
default left-to-right approach. SlidingWindowShort and
SlidingWindowShortRTL incorporate a "shortening" heuristic that
sometimes cuts windows short. The Z parameter restricts the
maximum number of zeroes that may appear in a window. If a window
is maximum length, contains at least one zero, and the bit
following the window is a one, then the window is shortened in
order to yield all trailing ones to the next window.

This new behavior was inspired by the windowing technique used for
the upper half of smooth isogeny primes in [isogenychains]. This
update also adds p512-2 from [isogenychains] to the result set.

Improvements with the new Ensemble:
- p256_scalar improved from +2 to -1
- p384_scalar improved from +1 to +0
- isop512_field (new) is -3
- p2519_field improved from 263 to 261

Notably, the isop512_field results are better than [isogenychains]
when using their weighting metric (square = 0.8 * multiply).