internal/bigvector: use interface to allocations for basis vectors (#106)
By far the largest number of allocations comes from the creation of bigvector
basis vectors. These just contain zeros and ones. This PR changes the Vector
type to an interface so that the basis type can be implemented without any
allocations at all.

Unfortunately, this does complicate the interface, since it requires the user
to not write to any integers returned from Idx(). Consider this okay since
it's an internal package.

Updates #60
Updates #25